A basic parser is available
I'm not terribly satisfied with this parser, but for now it can
understand some sentences, with keyword capture and optional fragments.

It would probably be wise to abandon building a single regular
expression and, instead, opt for a more flexible parser that tries to
match "manually" as much as possible.

In any case, it seems clear that "incomplete" cases will have to be
handled by each sentence handler.